moved common query builder code to common package.
