+++ src/main/java/com/hashedin/redmask/factory/DataMasking.java
@@ -2,11 +2,13 @@ package com.hashedin.redmask.factory;
 
 import java.io.IOException;
 
+import com.hashedin.redmask.exception.RedmaskConfigException;
+
 public abstract class DataMasking {
 
-  // Validate credentials, table, column etc config.
+  // TODO :Abstract validate credentials, table, column etc config.
   
-  public abstract void generateSqlQueryForMasking();
+  public abstract void generateSqlQueryForMasking() throws RedmaskConfigException;
   
   public abstract void executeSqlQueryForMasking() throws IOException, ClassNotFoundException;
