+++ src/main/java/com/hashedin/redmask/function/BigIntRangeMasking.java
@@ -1,4 +1,4 @@
-package com.hashedin.redmask.postgres.function;
+package com.hashedin.redmask.function;
 
 import com.hashedin.redmask.common.MaskingQueryUtil;
 import com.hashedin.redmask.common.MaskingRuleDef;
@@ -9,6 +9,7 @@ import com.hashedin.redmask.exception.RedmaskConfigException;
 import com.hashedin.redmask.exception.RedmaskRuntimeException;
 
 import freemarker.template.TemplateException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -19,25 +20,25 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * This masking function convert a column of type numeric into a range of numeric,
+ * This masking function convert a column of type bigint into a range of bigint,
  * with the range equal to the step parameter.
  */
-public class NumericRangeMasking extends MaskingRuleDef {
+public class BigIntRangeMasking extends MaskingRuleDef {
 
-  private static final Logger log = LoggerFactory.getLogger(NumericRangeMasking.class);
+  private static final Logger log = LoggerFactory.getLogger(BigIntRangeMasking.class);
 
   private static final String PARAM_STEP = "step";
 
   private static final String PARAM_STEP_DEFAULT = "10";
 
-  public NumericRangeMasking(
+  public BigIntRangeMasking(
       String columnName,
       MaskType maskType,
       Map<String, String> maskParams) {
     super(columnName, maskType, maskParams);
   }
 
-  public NumericRangeMasking() {
+  public BigIntRangeMasking() {
   }
 
   /**
@@ -50,8 +51,7 @@ public class NumericRangeMasking extends MaskingRuleDef {
   public void addFunctionDefinition(TemplateConfiguration config, Set<String> funcSet,
                                     String dbType) {
     try {
-      funcSet.add(MaskingQueryUtil.maskIntegerRange(config, dbType));
-      funcSet.add(MaskingQueryUtil.maskNumericRange(config, dbType));
+      funcSet.add(MaskingQueryUtil.maskBigIntRange(config, dbType));
       log.info("Function added for Mask Type {}", this.getMaskType());
     } catch (IOException | TemplateException ex) {
       throw new RedmaskRuntimeException(String.format("Error occurred while adding MaskFunction"
@@ -73,12 +73,14 @@ public class NumericRangeMasking extends MaskingRuleDef {
   @Override
   public String getSubQuery(TemplateConfiguration config, String tableName)
       throws RedmaskConfigException {
+
     List<String> paramsList = new ArrayList<>();
     paramsList.add(this.getColumnName());
+
     try {
       if (validateAndAddParameters(paramsList)) {
         return MaskingQueryUtil.processQueryTemplate(config,
-            MaskingConstants.MASK_NUMERIC_RANGE_FUNC, paramsList);
+            MaskingConstants.MASK_BIGINT_RANGE_FUNC, paramsList);
       }
     } catch (IOException | TemplateException ex) {
       throw new RedmaskRuntimeException(String.format("Error occurred while making SQL Sub query"
@@ -98,12 +100,14 @@ public class NumericRangeMasking extends MaskingRuleDef {
    * The Function will add the default value of the parameters value is not passed in the
    * maskparams config.
    * </p>
+   *
    * @param parameters List of parameters required to create the intended mask.
    * @return The list of validated parameter
    * @throws RedmaskConfigException
    */
-  private boolean validateAndAddParameters(List<String> parameters)
+  protected boolean validateAndAddParameters(List<String> parameters)
       throws RedmaskConfigException {
+
     for (String key : this.getMaskParams().keySet()) {
       if (!key.equals(PARAM_STEP)) {
         throw new RedmaskConfigException("Unrecognised parameter" + key + " supplied to "
