+++ src/main/java/com/hashedin/redmask/function/EmailMasking.java
@@ -1,4 +1,4 @@
-package com.hashedin.redmask.postgres.function;
+package com.hashedin.redmask.function;
 
 import com.hashedin.redmask.common.MaskingQueryUtil;
 import com.hashedin.redmask.common.MaskingRuleDef;
@@ -9,6 +9,7 @@ import com.hashedin.redmask.exception.RedmaskConfigException;
 import com.hashedin.redmask.exception.RedmaskRuntimeException;
 
 import freemarker.template.TemplateException;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -19,25 +20,46 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * This masking function convert a column of type integer into a range of integer,
- * with the range equal to the step parameter.
+ * This masking function is used to mask column containing email data entered as string.
+ * <p>
+ *   This function have the following variations :-
+ *   <p>
+ *     EMAIL_SHOW_DOMAIN : This shows only the domain part of the email address.
+ *   </p>
+ *   <p>
+ *     EMAIL_SHOW_FIRST_CHARACTER_DOMAIN : This shows the first character and the domain part of the
+ *     email address.
+ *   </p>
+ *   <p>
+ *     EMAIL_SHOW_FIRST_CHARACTERS : This shows the first few character equal to number passed in
+ *     the show_first parameter.
+ *   </p>
+ *   <p>
+ *     EMAIL_MASK_ALPHANUMERIC : This mask all the alphanumeric character in the email address.
+ *   </p>
+ * </p>
  */
-public class IntegerRangeMasking extends MaskingRuleDef {
+public class EmailMasking extends MaskingRuleDef {
 
-  private static final Logger log = LoggerFactory.getLogger(IntegerRangeMasking.class);
+  private static final Logger log = LoggerFactory.getLogger(EmailMasking.class);
 
-  private static final String PARAM_STEP = "step";
+  private static final String MASK_TYPE_SHOW_DOMAIN = "'domain'";
+  private static final String MASK_TYPE_SHOW_FIRST_DOMAIN = "'firstndomain'";
+  private static final String MASK_TYPE_SHOW_FIRST_CHAR = "'firstN'";
+  private static final String MASK_TYPE_SHOW_SPECIAL_CHAR = "'nonspecialcharacter'";
 
-  private static final String PARAM_STEP_DEFAULT = "10";
+  private static final String PARAM_SHOW_FIRST_CHARACTERS = "show_first";
+  private static final String PARAM_SHOW_FIRST_CHARACTERS_DEFAULT = "0";
 
-  public IntegerRangeMasking(
+
+  public EmailMasking(
       String columnName,
       MaskType maskType,
       Map<String, String> maskParams) {
     super(columnName, maskType, maskParams);
   }
 
-  public IntegerRangeMasking() {
+  public EmailMasking() {
   }
 
   /**
@@ -50,7 +72,8 @@ public class IntegerRangeMasking extends MaskingRuleDef {
   public void addFunctionDefinition(TemplateConfiguration config, Set<String> funcSet,
                                     String dbType) {
     try {
-      funcSet.add(MaskingQueryUtil.maskIntegerRange(config, dbType));
+      funcSet.add(MaskingQueryUtil.maskString(config, dbType));
+      funcSet.add(MaskingQueryUtil.maskEmail(config, dbType));
       log.info("Function added for Mask Type {}", this.getMaskType());
     } catch (IOException | TemplateException ex) {
       throw new RedmaskRuntimeException(String.format("Error occurred while adding MaskFunction"
@@ -77,15 +100,15 @@ public class IntegerRangeMasking extends MaskingRuleDef {
     try {
       if (validateAndAddParameters(paramsList)) {
         return MaskingQueryUtil.processQueryTemplate(config,
-            MaskingConstants.MASK_INTEGER_RANGE_FUNC, paramsList);
+            MaskingConstants.MASK_EMAIL_FUNC, paramsList);
       }
     } catch (IOException | TemplateException ex) {
+
       throw new RedmaskRuntimeException(String.format("Error occurred while making SQL Sub query"
               + "for column  %s  in table %s for Mask Type %s ", this.getColumnName(),
           tableName, this.getMaskType()), ex);
     }
     return this.getColumnName();
-
   }
 
   /**
@@ -98,7 +121,6 @@ public class IntegerRangeMasking extends MaskingRuleDef {
    * The Function will add the default value of the parameters value is not passed in the
    * maskparams config.
    * </p>
-   *
    * @param parameters List of parameters required to create the intended mask.
    * @return The list of validated parameter
    * @throws RedmaskConfigException
@@ -106,22 +128,42 @@ public class IntegerRangeMasking extends MaskingRuleDef {
   private boolean validateAndAddParameters(List<String> parameters)
       throws RedmaskConfigException {
     for (String key : this.getMaskParams().keySet()) {
-      if (!key.equals(PARAM_STEP)) {
+      if (!key.equals(PARAM_SHOW_FIRST_CHARACTERS)) {
         throw new RedmaskConfigException("Unrecognised parameter" + key + " supplied to "
             + this.getMaskType() + " for column " + this.getColumnName());
       }
     }
-    if (this.getMaskParams().isEmpty() || this.getMaskParams() == null) {
-      parameters.add(PARAM_STEP);
-      return true;
+    switch (this.getMaskType()) {
+      case EMAIL_SHOW_DOMAIN:
+        parameters.add(MASK_TYPE_SHOW_DOMAIN);
+        break;
+
+      case EMAIL_SHOW_FIRST_CHARACTER_DOMAIN:
+        parameters.add(MASK_TYPE_SHOW_FIRST_DOMAIN);
+        break;
+
+      case EMAIL_SHOW_FIRST_CHARACTERS:
+        parameters.add(MASK_TYPE_SHOW_FIRST_CHAR);
+
+        if (this.getMaskParams().isEmpty() || this.getMaskParams() == null) {
+          parameters.add(PARAM_SHOW_FIRST_CHARACTERS_DEFAULT);
+        }
+        int showCharacters = Integer.parseInt(this.getMaskParams()
+            .getOrDefault(PARAM_SHOW_FIRST_CHARACTERS, PARAM_SHOW_FIRST_CHARACTERS_DEFAULT));
+        if (showCharacters > 0) {
+          parameters.add(String.valueOf(showCharacters));
+        } else {
+          throw new RedmaskConfigException(
+              String.format("\'%s\' value should be greater than 0",
+                  PARAM_SHOW_FIRST_CHARACTERS));
+        }
+        break;
+      case EMAIL_MASK_ALPHANUMERIC:
+        parameters.add(MASK_TYPE_SHOW_SPECIAL_CHAR);
+        break;
+      default:
+        return false;
     }
-    int step = Integer.parseInt(this.getMaskParams().getOrDefault(PARAM_STEP, PARAM_STEP_DEFAULT));
-    if (step > 0) {
-      parameters.add(String.valueOf(step));
-      return true;
-    } else {
-      throw new RedmaskConfigException(
-          String.format("\'%s\' value should be greater than 0", PARAM_STEP));
-    }
+    return true;
   }
 }
