+++ src/main/java/com/hashedin/redmask/function/FixedValueIntegerMasking.java
@@ -1,4 +1,4 @@
-package com.hashedin.redmask.postgres.function;
+package com.hashedin.redmask.function;
 
 import com.hashedin.redmask.common.MaskingQueryUtil;
 import com.hashedin.redmask.common.MaskingRuleDef;
@@ -14,36 +14,30 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
 /**
- * This masking function masks an integer type column by a random number between the min
- * and max parameters.
+ * This masking function mask a integer type column by the fixed number passed as the
+ * value parameter or the default value 0.
  */
-public class InRangeIntegerMasking extends MaskingRuleDef {
+public class FixedValueIntegerMasking extends MaskingRuleDef {
 
-  private static final Logger log = LoggerFactory.getLogger(InRangeIntegerMasking.class);
+  private static final Logger log = LoggerFactory.getLogger(FixedValueIntegerMasking.class);
 
-  private static final String PARAM_MINIMUM = "min";
-  private static final String PARAM_MINIMUM__DEFAULT = "0";
+  private static final String PARAM_VALUE = "value";
 
-  private static final String PARAM_MAXIMUM = "max";
-  private static final String PARAM_MAXIMUM__DEFAULT = "10";
+  private static final String PARAM_VALUE_DEFAULT = "0";
 
-  private static final List<String> EXPECTED_PARAMETERS_LIST = new ArrayList<String>(
-      Arrays.asList(PARAM_MAXIMUM, PARAM_MINIMUM));
-
-  public InRangeIntegerMasking(
+  public FixedValueIntegerMasking(
       String columnName,
       MaskType maskType,
       Map<String, String> maskParams) {
     super(columnName, maskType, maskParams);
   }
 
-  public InRangeIntegerMasking() {
+  public FixedValueIntegerMasking() {
   }
 
   /**
@@ -56,7 +50,7 @@ public class InRangeIntegerMasking extends MaskingRuleDef {
   public void addFunctionDefinition(TemplateConfiguration config, Set<String> funcSet,
                                     String dbType) {
     try {
-      funcSet.add(MaskingQueryUtil.maskIntegerInRange(config, dbType));
+      funcSet.add(MaskingQueryUtil.maskIntegerFixedValue(config, dbType));
       log.info("Function added for Mask Type {}", this.getMaskType());
     } catch (IOException | TemplateException ex) {
       throw new RedmaskRuntimeException(String.format("Error occurred while adding MaskFunction"
@@ -83,7 +77,7 @@ public class InRangeIntegerMasking extends MaskingRuleDef {
     try {
       if (validateAndAddParameters(paramsList)) {
         return MaskingQueryUtil.processQueryTemplate(config,
-            MaskingConstants.MASK_INTEGER_WITHIN_RANGE_FUNC, paramsList);
+            MaskingConstants.MASK_INTEGER_FIXED_VALUE_FUNC, paramsList);
       }
     } catch (IOException | TemplateException ex) {
       throw new RedmaskRuntimeException(String.format("Error occurred while making SQL Sub query"
@@ -111,27 +105,18 @@ public class InRangeIntegerMasking extends MaskingRuleDef {
   private boolean validateAndAddParameters(List<String> parameters)
       throws RedmaskConfigException {
     for (String key : this.getMaskParams().keySet()) {
-      if (!EXPECTED_PARAMETERS_LIST.contains(key)) {
+      if (!key.equals(PARAM_VALUE)) {
         throw new RedmaskConfigException("Unrecognised parameter" + key + " supplied to "
             + this.getMaskType() + " for column " + this.getColumnName());
       }
     }
-
     if (this.getMaskParams().isEmpty() || this.getMaskParams() == null) {
-      parameters.addAll(Arrays.asList(PARAM_MINIMUM__DEFAULT, PARAM_MAXIMUM__DEFAULT));
-    }
-
-    int min = Integer.parseInt(this.getMaskParams()
-        .getOrDefault(PARAM_MINIMUM, PARAM_MINIMUM__DEFAULT));
-    int max = Integer.parseInt(this.getMaskParams()
-        .getOrDefault(PARAM_MAXIMUM, PARAM_MAXIMUM__DEFAULT));
-    if (max > min) {
-      parameters.add(String.valueOf(min));
-      parameters.add(String.valueOf(max));
+      parameters.add(PARAM_VALUE);
       return true;
-    } else {
-      throw new RedmaskConfigException(
-          String.format("\'%s\' should be greater than \'%s\'", PARAM_MAXIMUM, PARAM_MINIMUM));
     }
+    int value = Integer.parseInt(this.getMaskParams()
+        .getOrDefault(PARAM_VALUE, PARAM_VALUE_DEFAULT));
+    parameters.add(String.valueOf(value));
+    return true;
   }
 }
