+++ src/main/java/com/hashedin/redmask/function/NumericRangeMasking.java
@@ -1,4 +1,4 @@
-package com.hashedin.redmask.postgres.function;
+package com.hashedin.redmask.function;
 
 import com.hashedin.redmask.common.MaskingQueryUtil;
 import com.hashedin.redmask.common.MaskingRuleDef;
@@ -19,26 +19,25 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * This masking function mask a float type column by the fixed number passed as the
- * value parameter or the default value of 0.00.
+ * This masking function convert a column of type numeric into a range of numeric,
+ * with the range equal to the step parameter.
  */
-public class FixedValueFloatMasking extends MaskingRuleDef {
+public class NumericRangeMasking extends MaskingRuleDef {
 
-  private static final Logger log = LoggerFactory.getLogger(FixedValueFloatMasking.class);
+  private static final Logger log = LoggerFactory.getLogger(NumericRangeMasking.class);
 
-  private static final String PARAM_VALUE = "value";
+  private static final String PARAM_STEP = "step";
 
-  private static final String PARAM_VALUE_DEFAULT = "0.00";
+  private static final String PARAM_STEP_DEFAULT = "10";
 
-
-  public FixedValueFloatMasking(
+  public NumericRangeMasking(
       String columnName,
       MaskType maskType,
       Map<String, String> maskParams) {
     super(columnName, maskType, maskParams);
   }
 
-  public FixedValueFloatMasking() {
+  public NumericRangeMasking() {
   }
 
   /**
@@ -51,7 +50,8 @@ public class FixedValueFloatMasking extends MaskingRuleDef {
   public void addFunctionDefinition(TemplateConfiguration config, Set<String> funcSet,
                                     String dbType) {
     try {
-      funcSet.add(MaskingQueryUtil.maskFloatFixedValue(config, dbType));
+      funcSet.add(MaskingQueryUtil.maskIntegerRange(config, dbType));
+      funcSet.add(MaskingQueryUtil.maskNumericRange(config, dbType));
       log.info("Function added for Mask Type {}", this.getMaskType());
     } catch (IOException | TemplateException ex) {
       throw new RedmaskRuntimeException(String.format("Error occurred while adding MaskFunction"
@@ -76,9 +76,9 @@ public class FixedValueFloatMasking extends MaskingRuleDef {
     List<String> paramsList = new ArrayList<>();
     paramsList.add(this.getColumnName());
     try {
-      if (this.validateAndAddParameters(paramsList)) {
+      if (validateAndAddParameters(paramsList)) {
         return MaskingQueryUtil.processQueryTemplate(config,
-            MaskingConstants.MASK_FLOAT_FIXED_VALUE_FUNC, paramsList);
+            MaskingConstants.MASK_NUMERIC_RANGE_FUNC, paramsList);
       }
     } catch (IOException | TemplateException ex) {
       throw new RedmaskRuntimeException(String.format("Error occurred while making SQL Sub query"
@@ -98,7 +98,6 @@ public class FixedValueFloatMasking extends MaskingRuleDef {
    * The Function will add the default value of the parameters value is not passed in the
    * maskparams config.
    * </p>
-   *
    * @param parameters List of parameters required to create the intended mask.
    * @return The list of validated parameter
    * @throws RedmaskConfigException
@@ -106,18 +105,22 @@ public class FixedValueFloatMasking extends MaskingRuleDef {
   private boolean validateAndAddParameters(List<String> parameters)
       throws RedmaskConfigException {
     for (String key : this.getMaskParams().keySet()) {
-      if (!key.equals(PARAM_VALUE)) {
+      if (!key.equals(PARAM_STEP)) {
         throw new RedmaskConfigException("Unrecognised parameter" + key + " supplied to "
             + this.getMaskType() + " for column " + this.getColumnName());
       }
     }
     if (this.getMaskParams().isEmpty() || this.getMaskParams() == null) {
-      parameters.add(PARAM_VALUE);
+      parameters.add(PARAM_STEP);
       return true;
     }
-    float value = Float.parseFloat(this.getMaskParams()
-        .getOrDefault(PARAM_VALUE, PARAM_VALUE_DEFAULT));
-    parameters.add(String.valueOf(value));
-    return true;
+    int step = Integer.parseInt(this.getMaskParams().getOrDefault(PARAM_STEP, PARAM_STEP_DEFAULT));
+    if (step > 0) {
+      parameters.add(String.valueOf(step));
+      return true;
+    } else {
+      throw new RedmaskConfigException(
+          String.format("\'%s\' value should be greater than 0", PARAM_STEP));
+    }
   }
 }
